backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
langgraph
requests
python-dotenv
openai
